Rename dao.ts to db.ts
